feat: add react router and vitest
