+++ .eslintrc.cjs
@@ -18,5 +18,6 @@ module.exports = {
     ],
     'react-compiler/react-compiler': 'error',
     'react-hooks/exhaustive-deps': 'off',
+    'prettier/prettier': 'warn',
   },
 }
